[FIX] Fix hang in sSafetyRead state in frame creation fsm
The frame creation state machine could hang in sSafetyRead state
if the frame buffer is read completely (iReadBuffDone). Then the
frame manipulator reported a frame buffer overflow (0x20) in the
status register.
This fix ensures that the state machine exits to sCrc in order to
create a frame with valid CRC.

Change-Id: I4909d9055d0989349f8e9a40388cd255e3a6f92a